requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
langchain==0.1.0
langchain-community==0.0.10
langchain-core>=0.1.16,<0.2
langchain-openai==0.0.5
python-dotenv==1.0.0
chromadb==0.4.22
sentence-transformers==2.2.2
huggingface-hub==0.16.4
openai>=1.10.0,<2.0.0 